refactor: big refactory of rest client integration test classes using inheritance to solve duplicated code issues
