--- src/test/java/org/persapiens/account/restclient/DebitAccountRestClientFactory.java
@@ -1,45 +1,0 @@
-package org.persapiens.account.restclient;
-
-import java.util.Optional;
-
-import lombok.Data;
-import lombok.experimental.SuperBuilder;
-import org.persapiens.account.dto.CategoryDTO;
-import org.persapiens.account.dto.DebitAccountDTO;
-
-@SuperBuilder
-@Data
-public class DebitAccountRestClientFactory {
-
-	private String protocol;
-
-	private String servername;
-
-	private int port;
-
-	private CategoryRestClientFactory categoryRestClientFactory;
-
-	public DebitAccountRestClient debitAccountRestClient() {
-		return DebitAccountRestClient.builder()
-			.restClientHelper(RestClientHelper.<DebitAccountDTO>builder()
-				.endpoint("debitAccount")
-				.protocol(this.protocol)
-				.servername(this.servername)
-				.port(this.port)
-				.build())
-			.build();
-	}
-
-	public DebitAccountDTO debitAccount(String description, String categoryDescription) {
-		Optional<DebitAccountDTO> findByDescription = debitAccountRestClient().findByDescription(description);
-		if (findByDescription.isEmpty()) {
-			CategoryDTO category = this.categoryRestClientFactory.category(categoryDescription);
-			DebitAccountDTO result = DebitAccountDTO.builder().description(description).category(category).build();
-			return debitAccountRestClient().save(result);
-		}
-		else {
-			return findByDescription.get();
-		}
-	}
-
-}
