+++ src/test/java/org/persapiens/account/restclient/DebitAccountRestClientIT.java
@@ -6,7 +6,6 @@ import org.persapiens.account.AccountApplication;
 import org.persapiens.account.common.CategoryConstants;
 import org.persapiens.account.dto.DebitAccountDTO;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -14,31 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = AccountApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class DebitAccountRestClientIT {
-
-	private final String protocol = "http";
-
-	private final String servername = "localhost";
-
-	@Value("${local.server.port}")
-	private int port;
-
-	private CategoryRestClientFactory categoryRestClientFactory() {
-		return CategoryRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private DebitAccountRestClient debitAccountRestClient() {
-		return DebitAccountRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build()
-			.debitAccountRestClient();
-	}
+public class DebitAccountRestClientIT extends RestClientIT {
 
 	@Test
 	public void saveOne() {
@@ -47,7 +22,7 @@ public class DebitAccountRestClientIT {
 
 		DebitAccountDTO debitAccount = DebitAccountDTO.builder()
 			.description(description)
-			.category(categoryRestClientFactory().category(categoryDescription))
+			.category(category(categoryDescription))
 			.build();
 
 		// verify save operation
