+++ src/test/java/org/persapiens/account/restclient/EntryRestClientIT.java
@@ -12,7 +12,6 @@ import org.persapiens.account.common.EquityAccountConstants;
 import org.persapiens.account.common.OwnerConstants;
 import org.persapiens.account.dto.EntryDTO;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -20,57 +19,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = AccountApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class EntryRestClientIT {
-
-	private final String protocol = "http";
-
-	private final String servername = "localhost";
-
-	@Value("${local.server.port}")
-	private int port;
-
-	private EntryRestClient entryRestClient() {
-		return EntryRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build()
-			.entryRestClient();
-	}
-
-	private OwnerRestClientFactory ownerRestClientFactory() {
-		return OwnerRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private CategoryRestClientFactory categoryRestClientFactory() {
-		return CategoryRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private EquityAccountRestClientFactory equityAccountRestClientFactory() {
-		return EquityAccountRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.categoryRestClientFactory(categoryRestClientFactory())
-			.build();
-	}
-
-	private CreditAccountRestClientFactory creditAccountRestClientFactory() {
-		return CreditAccountRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.categoryRestClientFactory(categoryRestClientFactory())
-			.build();
-	}
+public class EntryRestClientIT extends RestClientIT {
 
 	@Test
 	public void saveOne() {
@@ -84,9 +33,9 @@ public class EntryRestClientIT {
 			.value(new BigDecimal(543))
 			.date(LocalDateTime.now())
 			.note("saving the internship")
-			.owner(ownerRestClientFactory().owner(mother))
-			.inAccount(equityAccountRestClientFactory().equityAccount(savings, bank))
-			.outAccount(creditAccountRestClientFactory().creditAccount(internship, salary))
+			.owner(owner(mother))
+			.inAccount(equityAccount(savings, bank))
+			.outAccount(creditAccount(internship, salary))
 			.build();
 
 		// verify save operation
