+++ src/test/java/org/persapiens/account/restclient/EquityAccountRestClientIT.java
@@ -6,7 +6,6 @@ import org.persapiens.account.AccountApplication;
 import org.persapiens.account.common.CategoryConstants;
 import org.persapiens.account.dto.EquityAccountDTO;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -14,31 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = AccountApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class EquityAccountRestClientIT {
-
-	private final String protocol = "http";
-
-	private final String servername = "localhost";
-
-	@Value("${local.server.port}")
-	private int port;
-
-	private CategoryRestClientFactory categoryRestClientFactory() {
-		return CategoryRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private EquityAccountRestClient equityAccountRestClient() {
-		return EquityAccountRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build()
-			.equityAccountRestClient();
-	}
+public class EquityAccountRestClientIT extends RestClientIT {
 
 	@Test
 	public void saveOne() {
@@ -47,7 +22,7 @@ public class EquityAccountRestClientIT {
 
 		EquityAccountDTO equityAccount = EquityAccountDTO.builder()
 			.description(description)
-			.category(categoryRestClientFactory().category(categoryDescription))
+			.category(category(categoryDescription))
 			.build();
 
 		// verify save operation
