+++ src/test/java/org/persapiens/account/restclient/OwnerEquityAccountInitialValueRestClientIT.java
@@ -10,7 +10,6 @@ import org.persapiens.account.common.EquityAccountConstants;
 import org.persapiens.account.common.OwnerConstants;
 import org.persapiens.account.dto.OwnerEquityAccountInitialValueDTO;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -18,48 +17,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = AccountApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class OwnerEquityAccountInitialValueRestClientIT {
-
-	private final String protocol = "http";
-
-	private final String servername = "localhost";
-
-	@Value("${local.server.port}")
-	private int port;
-
-	private OwnerEquityAccountInitialValueRestClient ownerEquityAccountInitialValueRestClient() {
-		return OwnerEquityAccountInitialValueRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build()
-			.ownerEquityAccountInitialValueRestClient();
-	}
-
-	private OwnerRestClientFactory ownerRestClientFactory() {
-		return OwnerRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private CategoryRestClientFactory categoryRestClientFactory() {
-		return CategoryRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.build();
-	}
-
-	private EquityAccountRestClientFactory equityAccountRestClientFactory() {
-		return EquityAccountRestClientFactory.builder()
-			.protocol(this.protocol)
-			.servername(this.servername)
-			.port(this.port)
-			.categoryRestClientFactory(categoryRestClientFactory())
-			.build();
-	}
+public class OwnerEquityAccountInitialValueRestClientIT extends RestClientIT {
 
 	@Test
 	public void saveOne() {
@@ -69,8 +27,8 @@ public class OwnerEquityAccountInitialValueRestClientIT {
 
 		OwnerEquityAccountInitialValueDTO ownerEquityAccountInitialValue = OwnerEquityAccountInitialValueDTO.builder()
 			.value(new BigDecimal(1000))
-			.owner(ownerRestClientFactory().owner(mother))
-			.equityAccount(equityAccountRestClientFactory().equityAccount(savings, bank))
+			.owner(owner(mother))
+			.equityAccount(equityAccount(savings, bank))
 			.build();
 
 		// verify save operation
